+++ core/src/thing/Thing.types.spec.ts
@@ -1,0 +1,64 @@
+import { graph, sym } from "rdflib";
+import { Thing } from "./Thing";
+
+describe("Thing", function () {
+  describe("types", () => {
+    it("are empty is nothing is found in store", () => {
+      const store = graph();
+      const it = new Thing("https://jane.doe.example/resource#it", store);
+      expect(it.types()).toEqual([]);
+    });
+
+    it("contains the single type of a resource", () => {
+      const store = graph();
+      const it = new Thing("https://jane.doe.example/resource#it", store);
+      store.add(
+        sym("https://jane.doe.example/resource#it"),
+        sym("http://www.w3.org/1999/02/22-rdf-syntax-ns#type"),
+        sym("https://vocab.example#TypeA")
+      );
+      expect(it.types()).toEqual(["https://vocab.example#TypeA"]);
+    });
+
+    it("contains all the types of a resource", () => {
+      const store = graph();
+      const it = new Thing("https://jane.doe.example/resource#it", store);
+      store.add(
+        sym("https://jane.doe.example/resource#it"),
+        sym("http://www.w3.org/1999/02/22-rdf-syntax-ns#type"),
+        sym("https://vocab.example#TypeA")
+      );
+      store.add(
+        sym("https://jane.doe.example/resource#it"),
+        sym("http://www.w3.org/1999/02/22-rdf-syntax-ns#type"),
+        sym("https://vocab.example#TypeB")
+      );
+      store.add(
+        sym("https://jane.doe.example/resource#it"),
+        sym("http://www.w3.org/1999/02/22-rdf-syntax-ns#type"),
+        sym("https://vocab.example#TypeC")
+      );
+      const types = it.types();
+      expect(types).toContain("https://vocab.example#TypeA");
+      expect(types).toContain("https://vocab.example#TypeB");
+      expect(types).toContain("https://vocab.example#TypeC");
+      expect(types).toHaveLength(3);
+    });
+
+    it("does not contain types of other things or other properties of the thing", () => {
+      const store = graph();
+      const it = new Thing("https://jane.doe.example/resource#it", store);
+      store.add(
+        sym("https://jane.doe.example/resource#other"),
+        sym("http://www.w3.org/1999/02/22-rdf-syntax-ns#type"),
+        sym("https://vocab.example#TypeA")
+      );
+      store.add(
+        sym("https://jane.doe.example/resource#it"),
+        sym("https://vocab.example#notAType"),
+        sym("https://vocab.example#TypeB")
+      );
+      expect(it.types()).toEqual([]);
+    });
+  });
+});